requirements.txt
amqp
asgiref
billiard
celery
click
click-didyoumean
click-plugins
click-repl
Django
django-redis
djangorestframework
drf-yasg
flower
humanize
inflection
kombu
packaging
prometheus_client
prompt_toolkit
psycopg2-binary
python-dateutil
python-dotenv
pytz
PyYAML
redis
setuptools
six
sqlparse
tornado
tzdata
uritemplate
vine
wcwidth
wheel
